Fixed a bug and wrap into new view
